\name{accmv-package}
\alias{accmv-package}
\alias{accmv}
\title{Inference with Available Complete-Case Missing Values}
\description{Implements inverse-probability-weighted, regression-adjustment,
and multiply-robust estimators under the available complete-case missing value
assumption of Cheng, Chen, Smith, and Zhao (2022).}
\references{Cheng G, Chen Y-C, Smith MA, Zhao Y-Q (2022).
\emph{Handling Nonmonotone Missing Data with Available Complete-Case Missing
Value Assumption}. \url{https://arxiv.org/abs/2207.02289}.}
\docType{package}
