\name{print.accmv_regression_result}
\alias{print.accmv_regression_result}
\title{Print an ACCMV Regression Result}
\usage{
\method{print}{accmv_regression_result}(x, ...)
}
\arguments{
\item{x}{An \code{accmv_regression_result} object.}
\item{...}{Unused.}
}
\value{The input object, invisibly.}
\description{Prints the ACCMV-weighted regression coefficients.}
